requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.3.1=pyhd8ed1ab_0
arrow-cpp=8.0.0=py310h3098874_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=38.0.1=py310h9ce1e76_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.0.90=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.78=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.0.76=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.0.90=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.0.90=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.0.90=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.7.1=pyhd8ed1ab_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
dill=0.3.6=pyhd8ed1ab_1
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.2=pyhd8ed1ab_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py310h5eee18b_0
fsspec=2022.11.0=pyhd8ed1ab_0
gds-tools=1.5.0.59=0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_0
huggingface_hub=0.11.1=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=py310h06a4308_0
importlib-metadata=5.1.0=pyha770c72_0
importlib_metadata=5.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py310h06a4308_0
ipython=8.4.0=py310h06a4308_0
jedi=0.18.1=py310h06a4308_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jupyter_client=7.2.2=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.5.0.59=0
libcufile-dev=1.5.0.59=0
libcurand=10.3.1.50=0
libcurand-dev=10.3.1.50=0
libcurl=7.86.0=h91b91d3_0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=he6d91bd_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.2=py310h5764c6d_1
multiprocess=0.70.12.2=py310h5764c6d_2
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
nsight-compute=2022.4.0.15=0
numexpr=2.8.4=py310h8879344_0
numpy=1.23.4=py310hd5efca6_0
numpy-base=1.23.4=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=22.0=pyhd8ed1ab_0
pandas=1.5.2=py310h1128e8f_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py310h468efa6_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=3.0.0=py310h5764c6d_1
python_abi=3.10=2_cp310
pytorch=1.13.0=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2022.6=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_4
pyzmq=23.2.0=py310h6a678d5_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py310h5eee18b_0
requests=2.28.1=py310h06a4308_0
responses=0.18.0=pyhd8ed1ab_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.1.1=py310h6a678d5_0
scipy=1.7.3=py310hfa59a62_0
setuptools=65.5.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hbd366e4_1
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py310h3dcd8bd_1
torchaudio=0.13.0=py310_cu116
torchvision=0.14.0=py310_cu116
tornado=6.1=py310h7f8727e_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.24.0=pyhd8ed1ab_0
typing-extensions=4.4.0=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py310h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=0.8.0=h7f98852_3
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py310h5764c6d_2
zeromq=4.3.4=h2531618_0
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
